backend/requirements.txt
fastapi
uvicorn
web3
python-dotenv
python-multipart